flask-celery-example/requirements.txt
flask
flask-restplus
celery
redis
python-dotenv
